+++ docs/changelog.md
@@ -13,6 +13,7 @@
 - Fix rendering issue with abnormal width of some symbols
 - Add ANSI OSC52 sequence support
 - Fix issue that disconnecting the session will directly close the tab instead of disconnecting the session
+- Improve the session manager page width can be freely adjusted
 
 ## [[V0.4.4](https://github.com/QQxiaoming/quardCRT/releases/tag/V0.4.4)] - 2024-03-26